+++ crates/types/src/lib.rs
@@ -75,6 +75,27 @@ impl SchemaType {
         })
     }
 
+    /// Create a nullable type for the provided schema. If already nullabe,
+    /// do nothing and return, otherwise convert to a union.
+    pub fn nullable(mut schema: SchemaType) -> SchemaType {
+        // Already a union
+        if let SchemaType::Union(inner) = &mut schema {
+            // Make nullable
+            if !inner
+                .variants_types
+                .iter()
+                .any(|t| matches!(**t, SchemaType::Null))
+            {
+                inner.variants_types.push(Box::new(SchemaType::Null));
+            }
+
+            return schema;
+        }
+
+        // Convert to a nullable union
+        SchemaType::union([schema, SchemaType::Null])
+    }
+
     /// Create an indexed/mapable object schema with the provided key and value types.
     pub fn object(key_type: SchemaType, value_type: SchemaType) -> SchemaType {
         SchemaType::Object(ObjectType {
@@ -250,6 +271,6 @@ impl<T: Schematic> Schematic for Box<T> {
 
 impl<T: Schematic> Schematic for Option<T> {
     fn generate_schema() -> SchemaType {
-        SchemaType::union_one([T::generate_schema(), SchemaType::Null])
+        SchemaType::nullable(T::generate_schema())
     }
 }